Move examples to tests folder
