--- examples/routers/item.py
@@ -1,8 +1,0 @@
-from simpleapi import Router
-
-router = Router()
-
-
-@router.get("/item")
-def item():
-    return {"name": "test", "price": 69}
