+++ tests/app.py
@@ -1,19 +1,23 @@
-from simpleapi import Request, Response, JSONResponse, SimpleAPI, custom_types
-from .routers import test
+from simpleapi import SimpleAPI, Request, Response, JSONResponse
+from .routers import item
+from pydantic import BaseModel
 
 app = SimpleAPI()
 
-app.add_router(prefix="/router", router=test.router)
+app.add_router(prefix="/router", router=item.router)
 
 
-# def current_user(f: custom_types.ViewFunction):
-#     """Middleware that adds user data to the request"""
+class Item(BaseModel):
+    name: str
+    price: float
 
-#     def decorator(request: Request):
-#         request.extra["user"] = request.body
-#         return f(request)
 
-#     return decorator
+items: list[Item] = []
+
+
+def current_user(request: Request):
+    """Middleware that adds user data to the request"""
+    request.extra["user"] = request.body
 
 
 @app.get("/hello")
@@ -31,18 +35,7 @@ def greet(request: Request):
 @app.post("/items")
 def post_item(request: Request):
     """Test post request returning body as JSON"""
-    print("request.body", request.body)
     return JSONResponse(body=request.body)
-
-
-@app.get("/html")
-def html():
-    """Test returning an HTML string"""
-    return Response(
-        code=200,
-        body="<html><body><h1>Hi</h1></body></html>",
-        content_type="text/html",
-    )
 
 
 @app.get("/400")
@@ -75,8 +68,7 @@ def options():
     return ""
 
 
-@app.post("/middleware")
-# @current_user
+@app.post("/middleware", middleware=[current_user])
 def middleware(request: Request):
     """Tests middleware"""
     return JSONResponse(body=request.extra["user"])
@@ -99,3 +91,104 @@ def dependency_injection_error(name: str, price: float, active: bool):
     """Test dependency injection"""
     # ? SimpleAPI should automatically return an error with this ever executing
     return {"name": name, "price": price}
+
+
+@app.post("/dependency-injection")
+def dep_injection(name: str, price: float):
+    """View function that uses depedency injection"""
+    return {"name": name, "price": price}
+
+
+@app.get("/1")
+def index():
+    """View function that takes no parameter and returns JSONResponse"""
+    return JSONResponse(body={"hello": "world"})
+
+
+@app.get("/2")
+def index2():
+    """View function that takes no parameter and returns a string"""
+    return "Hello, World!"
+
+
+@app.get("/3")
+def index3():
+    """View function that takes no parameter and returns an int"""
+    return 12
+
+
+@app.get("/4")
+def index4():
+    """View function that takes no parameter and returns a float"""
+    return 20.56
+
+
+@app.get("/5")
+def index5():
+    """View function that takes no parameter and returns a dict"""
+    person = {"name": "adhom", "age": 23}
+    return person
+
+
+@app.get("/6")
+def index6():
+    """View function that takes no parameter and returns a Pydantic BaseModel"""
+    item = Item(name="some item", price=2000)
+    return item
+
+
+@app.get("/greet/{first_name}/{last_name}")
+def greet_fullname(request: Request):
+    """Multiple dynamic route that greets users"""
+    fullname = request.params["first_name"] + " " + request.params["last_name"]
+    return JSONResponse(body={"fullname": fullname})
+
+
+@app.post("/pydantic-item")
+def post_pydantic_item(item: Item):
+    return item
+
+
+@app.get("/items")
+def get_item(request: Request):
+    q = request.query["q"][0]
+    results: list[dict] = []
+    for item in items:
+        if q in item.name:
+            results.append(item.dict())
+    return JSONResponse(body=results)
+
+
+@app.post("/image")
+def save_image(request: Request):
+    """
+    Receives an image and stores it on disk.
+    Assumes request content-type is multipart.
+    """
+    try:
+        image: bytes = request.form["image"]
+        with open("image", "wb") as file:
+            file.write(image)
+    except Exception as e:
+        print("Error while trying to save image")
+        print(str(e))
+        return JSONResponse(code=400, body={"body": "Invalid image"})
+    return Response(
+        body=image, content_type="image/*"
+    )  # Return the image as a response
+
+
+@app.get("/html")
+def html():
+    """Returns an HTML string"""
+    return Response(
+        code=200,
+        body="<html><body><h1>Hi</h1></body></html>",
+        content_type="text/html; charset=UTF-8",
+    )
+
+
+@app.get("/empty")
+def empty():
+    """Returns an empty response"""
+    return ""
